+++ examples/lib/stories/effects/effects.dart
@@ -7,6 +7,7 @@ import 'combined_effect.dart';
 import 'infinite_effect.dart';
 import 'move_effect.dart';
 import 'opacity_effect.dart';
+import 'remove_effect_example.dart';
 import 'rotate_effect.dart';
 import 'scale_effect.dart';
 import 'sequence_effect.dart';
@@ -67,5 +68,11 @@ void addEffectsStories(Dashbook dashbook) {
       'Color Effect',
       (_) => GameWidget(game: ColorEffectGame()),
       codeLink: baseLink('effects/color_effect.dart'),
+    )
+    ..add(
+      'Remove Effect',
+      (_) => GameWidget(game: RemoveEffectExample()),
+      codeLink: baseLink('effects/remove_effect_example.dart'),
+      info: RemoveEffectExample.description,
     );
 }